# openapi.yaml
openapi: 3.1.0
info:
  title: InvoiceParser API
  description: AI-powered invoice parsing API. Send a PDF or image, get clean structured JSON.
  version: 1.1.0
  contact:
    name: InvoiceParser Support

servers:
  - url: https://invoice-parser-api-gnmr.onrender.com
    description: Production

security:
  - ApiKeyAuth: []
  - BearerAuth: []

paths:
  /health:
    get:
      summary: Health check
      operationId: healthCheck
      security: []
      responses:
        "200":
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  version:
                    type: string
                    example: 1.0.0

  /v1/register:
    post:
      summary: Register for a free API key
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: you@company.com
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_key:
                    type: string
                    example: inv_a1b2c3d4e5f6...
                  plan:
                    type: string
                    example: free
                  max_calls:
                    type: integer
                    example: 100
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: email_exists
                  message:
                    type: string

  /v1/parse/invoice:
    post:
      summary: Parse a single invoice
      operationId: parseInvoice
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Invoice file (PDF, PNG, JPG, WebP, GIF). Max 10MB.
      responses:
        "200":
          description: Invoice parsed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/InvoiceData"
                  latency_ms:
                    type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"

  /v1/parse/url:
    post:
      summary: Parse an invoice from a URL
      operationId: parseInvoiceURL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                  description: URL pointing to a PDF or image file.
      responses:
        "200":
          description: Invoice parsed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/InvoiceData"
                  latency_ms:
                    type: integer
                  request_id:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"

  /v1/rotate-key:
    post:
      summary: Rotate API key
      operationId: rotateKey
      description: Generates a new API key and invalidates the old one.
      responses:
        "200":
          description: New API key generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_key:
                    type: string
                  message:
                    type: string
                  request_id:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"

  /v1/parse/batch:
    post:
      summary: Parse multiple invoices (Pro plan only)
      operationId: batchParseInvoice
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Up to 20 invoice files.
                webhook_url:
                  type: string
                  format: uri
                  description: If provided, results are POSTed here asynchronously.
      responses:
        "200":
          description: Batch processed synchronously
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  total:
                    type: integer
                  succeeded:
                    type: integer
                  failed:
                    type: integer
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/BatchResult"
                  latency_ms:
                    type: integer
        "202":
          description: Batch job queued (async with webhook)
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
                    example: pending
                  file_count:
                    type: integer
                  webhook_url:
                    type: string
        "403":
          description: Pro plan required

  /v1/parse/batch/{id}:
    get:
      summary: Get batch job status
      operationId: getBatchJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Batch job details
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
                    enum: [pending, processing, completed, failed]
                  file_count:
                    type: integer
                  completed:
                    type: integer
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/BatchResult"
        "404":
          description: Job not found

  /v1/usage:
    get:
      summary: Get API usage stats
      operationId: getUsage
      responses:
        "200":
          description: Usage statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    type: string
                  used_calls:
                    type: integer
                  max_calls:
                    type: integer
                  today_calls:
                    type: integer
                  month_calls:
                    type: integer

  /v1/dashboard:
    get:
      summary: Get full dashboard data
      operationId: getDashboard
      responses:
        "200":
          description: Dashboard data with daily breakdown and recent logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    type: string
                  used_calls:
                    type: integer
                  max_calls:
                    type: integer
                  today_calls:
                    type: integer
                  month_calls:
                    type: integer
                  daily_usage:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                        calls:
                          type: integer
                  recent_logs:
                    type: array
                    items:
                      type: object
                      properties:
                        endpoint:
                          type: string
                        status:
                          type: integer
                        latency_ms:
                          type: integer
                        created_at:
                          type: string
                  member_since:
                    type: string

  /v1/billing/checkout:
    post:
      summary: Create a Stripe checkout session
      operationId: createCheckout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan, success_url, cancel_url]
              properties:
                plan:
                  type: string
                  enum: [starter, pro]
                success_url:
                  type: string
                  format: uri
                cancel_url:
                  type: string
                  format: uri
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkout_url:
                    type: string
                    format: uri
        "503":
          description: Billing not configured
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string

  /metrics:
    get:
      summary: Prometheus metrics
      operationId: getMetrics
      security: []
      responses:
        "200":
          description: Prometheus text exposition format
          content:
            text/plain:
              schema:
                type: string

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer

  schemas:
    InvoiceData:
      type: object
      properties:
        invoice_number:
          type: string
        invoice_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        currency:
          type: string
        vendor:
          type: object
          properties:
            name:
              type: string
            address:
              type: string
            email:
              type: string
            phone:
              type: string
            tax_id:
              type: string
        customer:
          type: object
          properties:
            name:
              type: string
            address:
              type: string
            email:
              type: string
            phone:
              type: string
        line_items:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              quantity:
                type: number
              unit_price:
                type: number
              amount:
                type: number
        subtotal:
          type: number
        tax_rate:
          type: number
        tax_amount:
          type: number
        discount:
          type: number
        total:
          type: number
        payment_terms:
          type: string
        notes:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1

    BatchResult:
      type: object
      properties:
        index:
          type: integer
        filename:
          type: string
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/InvoiceData"
        error:
          type: string

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
              used_calls:
                type: integer
              max_calls:
                type: integer
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string